Propagate GTK "Kerning Format" dialog results

+++ fontforgeexe/gtk/c_context.h
@@ -47,8 +47,8 @@ typedef struct kerning_format_data {
     bool use_individual_pairs;
     bool guess_kerning_classes;
     double intra_class_dist;
-    double default_separation;
-    double min_kern;
+    int default_separation;
+    int min_kern;
     bool touching;
     bool kern_closer;
     bool autokern_new;
